Rename blueprint main files to __init__
