+++ application/blueprints/artist/__init__.py
